# apiDoc.json
{
    "swagger": "2.0",
    "info": {
        "title": "WU Researcher API Documentation",
        "version": "1.0.0",
        "description": "Documentation for your API"
    },
    "paths": {
...
        "/scopus/author": {
            "get": {
                "summary": "Get a list of authors",
                "tags": [
                    "Scopus Authors"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "sortField",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Field to sort by (h-index, document-count, name)"
                    },
                    {
                        "in": "query",
                        "name": "sortOrder",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Sorting order (asc, desc)"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Page number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuthorScopus"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/scopus/author/{scopus_id}": {
            "get": {
                "summary": "Get author by scopus_id",
                "tags": [
                    "Scopus Authors"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "scopus_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "scopus_id of the author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Author details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorScopus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Author not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/scopus/author/name/{authorName}": {
            "get": {
                "summary": "Get authors by name",
                "tags": [
                    "Scopus Authors"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "authorName",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Name of the author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors with matching name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuthorScopus"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/scopus/article/{eid}": {
            "get": {
                "summary": "Get an article by EID",
                "tags": [
                    "Scopus Articles"
                ],
                "description": "Retrieve an article by its EID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "eid",
                        "required": true,
                        "description": "EID of the article",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleScopus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "Article not found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/scopus/article/authorId/{scopus_id}": {
            "get": {
                "summary": "Get articles by scopus_id",
                "tags": [
                    "Scopus Articles"
                ],
                "description": "Retrieve articles by scopus ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "scopus_id",
                        "required": true,
                        "description": "Scopus ID of the author",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArticleScopus"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Articles not found for the given author ID",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "Articles not found"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/scopus/journal": {
            "get": {
                "summary": "Get a list of journals",
                "tags": [
                    "Journal"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "sortField",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Field to sort by (e.g. \"journal-name\")"
                    },
                    {
                        "in": "query",
                        "name": "sortOrder",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Sort order (\"asc\" or \"desc\")"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Page number for pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a list of journals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Journal"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/scopus/journal/{source_id}": {
            "get": {
                "summary": "Get a journal by source ID",
                "tags": [
                    "Journal"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "source_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Source ID of the journal"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with the journal data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Journal"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Journal not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "Journal not found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/scopus/journal/name/{journalName}": {
            "get": {
                "summary": "Search for journals by name",
                "tags": [
                    "Journal"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "journalName",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Name of the journal"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with matching journals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Journal"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ArticleScopus": {
                "type": "object",
                "properties": {
                    "eid": {
                        "type": "string"
                    },
                    "article_name": {
                        "type": "string"
                    },
                    "source_id": {
                        "type": "string"
                    },
                    "first_author": {
                        "type": "string"
                    },
                    "co_author": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "co_author_department": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "volume": {
                        "type": "string"
                    },
                    "issue": {
                        "type": "string"
                    },
                    "pages": {
                        "type": "string"
                    },
                    "document_type": {
                        "type": "string"
                    },
                    "source_type": {
                        "type": "string"
                    },
                    "issn": {
                        "type": "string"
                    },
                    "original_language": {
                        "type": "string"
                    },
                    "publisher": {
                        "type": "string"
                    },
                    "abstract": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "author_scopus_id": {
                        "type": "string"
                    }
                }
            },
            "AuthorScopus": {
                "type": "object",
                "properties": {
                    "author_scopus_id": {
                        "type": "string"
                    },
                    "author_name": {
                        "type": "string"
                    },
                    "citations": {
                        "type": "string"
                    },
                    "citations_by": {
                        "type": "string"
                    },
                    "documents": {
                        "type": "string"
                    },
                    "wu_documents": {
                        "type": "string"
                    },
                    "h_index": {
                        "type": "string"
                    },
                    "subject_area": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "citations_graph": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "year": {
                                    "type": "number"
                                },
                                "citations": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "documents_graph": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "year": {
                                    "type": "number"
                                },
                                "documents": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "url": {
                        "type": "string"
                    }
                }
            },
...
            "Journal": {
                "type": "object",
                "properties": {
                    "source_id": {
                        "type": "string"
                    },
                    "journal_name": {
                        "type": "string"
                    },
                    "scopus_coverage_years": {
                        "type": "string"
                    },
                    "publisher": {
                        "type": "string"
                    },
                    "issn": {
                        "type": "string"
                    },
                    "eissn": {
                        "type": "string"
                    },
                    "source_type": {
                        "type": "string"
                    },
                    "subject_area": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "changeJournal": {
                        "type": "object"
                    },
                    "cite_source": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    }
}